Started using virtualenv Python
vendor/vendor_compile.sh now installs the virtual environment and setups
the packages under it.  I've done my best to point all references to the
python executable to the virtual executable.

Also had to change the order of vj-misc.el and vj-python.el in init.el
so that virtualenv PATH is first.